requirements.txt
streamlit
joblib
pillow
pytesseract
beautifulsoup4
scikit-learn
supabase